--- packages/zwave-js/src/lib/node/mixins/index.ts
@@ -1,3 +1,0 @@
-import { FirmwareUpdateMixin } from "./01_FirmwareUpdate";
-
-export class ZWaveNodeMixins extends FirmwareUpdateMixin {}
